Extract NetworkManager parent_manager to a mixin
Extract the logic for a NetworkManager belonging to a CloudManager
parent out of the base NetworkManager class and move it to a mixin.
